requirements.txt
regex >= 2024.11.6
paho-mqtt >= 2.1.0
